requirements.txt
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pygame==2.5.2
PyGLM==2.7.0
numpy==1.26.0
Pillow==10.2.0
pywavefront==1.3.3